# openapi/fc_openapi.yaml
openapi: 3.0.1
info:
  title: Eclipse XFSC Federated Catalogue
  description: REST API of the XFSC catalogue
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: https://fc-server.xfsc.org
    description: some future test environment
components:
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: >-
        Forbidden. The user does not have the permission to execute this
        request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: >-
        May contain hints how to solve the error or indicate what went wrong at
        the server. Must not outline any information about the internal
        structure of the server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ClientError:
      description: >-
        May contain hints how to solve the error or indicate what was wrong in
        the request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: HTTP Conflict 409
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...
    Asset:
      type: object
      properties:
        assetHash:
          type: string
        id: #!! Added id / credentialSubject
          type: string
        status:
          $ref: '#/components/schemas/AssetStatus'
        issuer:
          type: string
        validatorDids:
          description: >-
            The credentialSubjects of the parties that validated (parts of) the
            credential.
          type: array
          items:
            type: string
        uploadDatetime:
          type: string
          format: date-time
          example: '2022-03-01T13:00:00Z'
          description: The time stamp (ISO8601) when the asset was uploaded.
        statusDatetime:
          type: string
          format: date-time
          example: '2022-05-11T15:30:00Z'
          description: >-
            The last time stamp (ISO8601) the status changed (for this
            Catalogue)
        contentType:
          type: string
          description: MIME content-type of the uploaded asset (e.g. application/ld+json, application/pdf)
          example: 'application/ld+json'
        fileSize:
          type: integer
          format: int64
          description: Size of the uploaded asset in bytes
        warnings:
          type: array
          description: >-
            Warnings issued during processing (e.g. triples removed due to reserved namespace filtering).
            The operation succeeded for all remaining data. Absent or empty when no warnings occurred.
          items:
            type: string
          example:
            - "2 triple(s) were removed from your upload because they use the reserved internal namespace
               <https://projects.eclipse.org/projects/technology.xfsc/federated-catalogue/meta#>.
               This namespace is reserved for internal catalogue use only and cannot be set by external clients."
    AssetStatus:
      type: string
      enum:
        - active
        - eol
        - deprecated
        - revoked
    AssetResult:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Asset'
        content:
          type: string
...
    ResultPage:
      type: object
      properties:
        totalCount:
          type: integer
    Assets:
      allOf:
        - $ref: '#/components/schemas/ResultPage'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/AssetResult'
...
  parameters:
    OffsetParam:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      required: false
      description: The number of items to skip before starting to collect the result set.
    LimitParam:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
      required: false
      description: The number of items to return.
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /assets:
    get:
      tags:
        - Assets
      summary: Get the list of asset metadata in the Catalogue
      operationId: readAssets
      security:
        - jwt: []
      parameters: #!! Added filter parameters for all(?) metadata fields
        - name: uploadTimerange
          in: query
          description: >-
            Filter for the time range when the asset was uploaded to the
            catalogue. The time range has to be specified as start time and
            end time as ISO8601 timestamp separated by a `/`.
          required: false
          example: 2022-03-01T13:00:00Z/2022-05-11T15:30:00Z
          schema:
            type: string
        - name: statusTimerange
          in: query
          description: >-
            Filter for the time range when the status of the asset was last
            changed in the catalogue. The time range has to be specified as
            start time and end time as ISO8601 timestamp separated by a `/`.
          required: false
          example: 2022-03-01T13:00:00Z/2022-05-11T15:30:00Z
          schema:
            type: string
        - name: issuers
          in: query
          description: >-
            Filter for the issuer of the credential. This is the unique ID of the
            Participant that has prepared the credential.
          required: false
          schema:
            type: array
            items:
              type: string
          explode: false
        - name: validators
          in: query
          description: >-
            Filter for a validator of the credential. This is the unique ID of the
            Participant that validated (part of) the credential.
          required: false
          schema:
            type: array
            items:
              type: string
          explode: false
        - name: statuses
          in: query
          description: Filter for the status of the asset.
          required: false
          schema:
            type: array
            items:
              $ref: '#/components/schemas/AssetStatus'
          explode: false
        - name: ids
          in: query
          description: Filter for the id/credentialSubject of the credential.
          required: false
          schema:
            type: array
            items:
              type: string
          explode: false
        - name: hashes
          in: query
          description: Filter for the hash of the asset.
          required: false
          schema:
            type: array
            items:
              type: string
          explode: false
        - name: withMeta
          in: query
          description: Whether to include asset metadata in the response.
          required: false
          schema:
            type: boolean
            default: true
        - name: withContent
          in: query
          description: Whether to include asset content in the response.
          required: false
          schema:
            type: boolean
            default: false
        - $ref: '#/components/parameters/OffsetParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of metadata of available assets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assets'
        '400':
          $ref: '#/components/responses/ClientError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags:
        - Assets
      summary: Add a new asset to the catalogue
      description: >-
        Upload an asset. RDF content (JSON-LD credentials) is automatically
        detected and routed through the verification pipeline. Non-RDF content
        is stored directly.
      operationId: addAsset
      security:
        - jwt: []
      requestBody:
        description: >-
          The asset content. For JSON-LD credentials, use application/json,
          application/vc+ld+json, or application/vp+ld+json. For JWT-wrapped
          Gaia-X credentials per ICAM 24.07 (Loire release), use application/vc+ld+json+jwt
          (VC) or application/vp+ld+json+jwt (VP). For non-RDF assets, use
          multipart/form-data or application/octet-stream (handled by the
          server outside the generated delegate pattern).
        content:
          application/json:
            schema:
              type: string
          application/ld+json:
            schema:
              type: string
          application/vc+ld+json:
            schema:
              type: string
          application/vp+ld+json:
            schema:
              type: string
          application/vc+ld+json+jwt:
            schema:
              type: string
          application/vp+ld+json+jwt:
            schema:
              type: string
        required: true
      responses:
        '201':
          description: Created
          headers:
            location:
              description: URL of the added asset.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '400':
          $ref: '#/components/responses/ClientError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/ServerError'
  /assets/{id}:
    get:
      tags:
        - Assets
      summary: Read an asset by its IRI.
      description: >
        Returns the content of a single RDF asset identified by its IRI.
        Non-RDF assets (PDF, images, binary files) are stored in the FileStore
        and cannot be retrieved via this endpoint — a 400 error is returned
        with the asset's hash and content-type for reference.
      operationId: readAssetById
      parameters:
        - name: id
          in: path
          description: IRI identifier of the asset (DID, URN, or HTTP IRI)
          required: true
          schema:
            type: string
          examples:
            did:
              summary: DID Web identifier
              value: 'did:web:example.com:participant'
            urn:
              summary: UUID URN (non-RDF assets)
              value: 'urn:uuid:550e8400-e29b-41d4-a716-446655440000'
            http:
              summary: HTTP IRI
              value: 'https://example.org/asset/123'
      security:
        - jwt: []
      responses:
        '200':
          description: The requested asset (RDF content)
          content:
            application/json:
              schema:
                type: string
        '400':
          description: >
            The asset exists but is a non-RDF asset (e.g., PDF, image, binary).
            Raw content download is not supported via this endpoint.
            The response includes the asset's hash and content-type.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /assets/{asset_hash}:
    delete:
      tags:
        - Assets
      summary: Completely delete an asset by its content hash
      description: >
        Unlike other asset endpoints which use IRI-based identification (/assets/{id}),
        the delete operation uses the SHA-256 content hash to guarantee unambiguous
        single-row targeting. See ADR 7 — Delete Operation Exception.
      operationId: deleteAsset
      security:
        - jwt: []
      parameters:
        - name: asset_hash
          in: path
          description: Hash identifier of the asset
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '400':
          $ref: '#/components/responses/ClientError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /assets/{asset_hash}/revoke:
    post:
      tags:
        - Assets
      summary: Change the lifecycle state of an asset to revoked
      description: >
        Like the delete operation, the revoke endpoint uses the SHA-256 content hash
        to guarantee unambiguous single-row targeting. Only ACTIVE assets can be revoked.
        When versioning is introduced (CAT-FR-LM-01), this endpoint will be replaced by
        version-specific revocation (POST /assets/{id}/versions/{v}/revoke).
        See ADR 7 — Delete Operation Exception.
      operationId: revokeAsset
      security:
        - jwt: []
      parameters:
        - name: asset_hash
          in: path
          description: Hash identifier of the asset
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Revoked
          headers:
            location:
              description: URL of the asset
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Asset'
        '400':
          $ref: '#/components/responses/ClientError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...